Add Ninja to regression tests
only one test needed fixed which was using add_custom_target to generate a file in a way that was not supported by ninja, this was not an issue relating to xcommon cmake so I updated the test to use add_custom_command. This works.

Remove ninja tests but leave infra

dont want to fix right now but I think the changes to allow changing
the generator easily will be useful if I get motivated in the future to
try and fix.

add back ninja generator

Revert accidental change